port SearchFilterDataExporter to ts
